fix #62

+++ Dan2Doozer/GetItemsFromDoozer.m
@@ -45,7 +45,7 @@
         NSDictionary *jsonDict = (NSDictionary *) responseObject;
         itemsArray = [jsonDict objectForKey:@"items"];
         //NSLog(@" heres' the server response =%@", itemsArray);
-        NSLog(@"Count of items from doozer server = %lu", itemsArray.count);
+        NSLog(@"Count of items from doozer server = %lu", (unsigned long)itemsArray.count);
         
         _completionHandler(itemsArray);
         _completionHandler = nil;
